Move return url func to utils
