+++ kobo/apps/stripe/utils.py
@@ -1,5 +1,6 @@
 from math import ceil, floor
 
+from django.conf import settings
 from djstripe.models import Price
 
 
@@ -16,3 +17,14 @@ def get_total_price_for_quantity(price: Price, quantity: int):
         else:
             total_price = floor(total_price)
     return total_price * price.unit_amount
+
+def generate_return_url(product_metadata):
+    """
+    Determine which frontend page Stripe should redirect users to
+    after they make a purchase or manage their account.
+    """
+    base_url = settings.KOBOFORM_URL + '/#/account/'
+    return_page = (
+        'addons' if product_metadata['product_type'] == 'addon' else 'plan'
+    )
+    return base_url + return_page
